requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.18.0
Pillow==6.2.1
protobuf==3.11.2
python-resize-image==1.1.19
requests==2.22.0
six==1.13.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.7
uWSGI==2.0.18
Werkzeug==0.16.0
wrapt==1.11.2
